fix: 공용 검색 필터 페이지 이동시에 필터 초기화 되도록 수정
